+++ NEWS.md
@@ -1,3 +1,4 @@
 # tsentiment 0.1.1
 
 * Fix a `Policy clause` for saving file.
+* Change `Workspace Global` to `env` for exapmple : APIinfo